requirements.txt
streamlit
pandas
numpy
fpdf2
pillow
openpyxl